# src/main/java/it/pagopa/atmlayer/wf/process/bean/TaskResponse.java
package it.pagopa.atmlayer.wf.process.bean;

import java.util.List;

import io.quarkus.runtime.annotations.RegisterForReflection;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@AllArgsConstructor
@NoArgsConstructor
@RegisterForReflection
public class TaskResponse {

    private List<Task> tasks;

    private String transactionId;

}